[SDS-335] Resolve import cycles using postponed annotations (#137)
* [SDS-335] Resolve import cycles using postponed annotations

Postponed annotations allow us to use forward code references without having access to them statically, but we can have access at runtime. Using the TYPE_CHECKING constant from typing makes sure that we don't have circular dependencies in the actual code.

The "correct" approach would be to extract interfaces. For example, we could have a generic API interface or a generic Job interface that exposes the methods and properties that we need access to. However, this is not particularly "Pythonese", so this approach seemed more lightweight.

* Address review comments

* Add a .coveragerc so we can ignore TYPE_CHECKING warning lines for coverage